restructured dialog components
