+++ frontend/src/components/Administration/Users/Dialog/DeleteUser.vue
