+++ frontend/src/components/Management/Dialog/CreatePlatformBinding.vue
