+++ frontend/src/components/Management/Dialog/CreatePlatformVersion.vue
