+++ frontend/src/components/Management/Dialog/DeletePlatformBinding.vue
