+++ frontend/src/components/Management/Dialog/DeletePlatformVersion.vue
