+++ frontend/src/components/common/Game/Dialog/Asset/DeleteAssets.vue
