+++ frontend/src/components/common/Game/Dialog/CopyDownloadLink.vue
