+++ frontend/src/components/common/Platform/Dialog/DeleteFirmware.vue
