+++ frontend/src/components/common/Platform/Dialog/DeletePlatform.vue
