+++ frontend/src/components/common/Platform/Dialog/UploadFirmware.vue
