+++ frontend/src/views/Administration.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import UsersTable from "@/components/Administration/UsersTable.vue";
+import UsersTable from "@/components/Administration/Users/Table.vue";
 import Tasks from "@/layouts/Administration/Tasks.vue";
 import storeAuth from "@/stores/auth";
